Wrap checkResults
This is in preparation of handling child processes that don't return.
In such a case, the default checkResults can't be called as the Promise
returned by childProcessData never resolves.

But this is still a valid case when doing IO with a running child
process.

So allow for a case where spawnTest doesn't return the usual promise:
Then a Muter is used to catch all messages until the test is over.

Of course the checkResults provided by the user must pole regularly
its results to actually be of any use.